Check for valid receiver device to detect a running transfer mode
Transferring() only returns true when source and destination device are different. There can be a transfer mode active where source and destination are the same. This can be detected by testing the receiver device of the transfer control class.

This fixes not working or stuttering audio in case of such a transfer mode.